Clear base cloner class
